Stop using json for serialization
Json files are very nice for some things but these messages are really
small and I don't want to force the c++ users to manually format the
json strings or have to import a json library. Instead, I'll force the
python users to pack their data into a struct. I think this will be
better for everyone.